Renamed the entry-point module
